+++ src/app/login/[[...login]]/page.tsx
@@ -4,22 +4,23 @@ import * as React from "react";
 import { useSignIn } from "@clerk/nextjs";
 import { useRouter } from "next/navigation";
 import { H1 } from "@/components/ui/H1";
-import { Input } from "@/components/ui/input";
-import Card from "@/components/ui/Card";
-import { Button } from "@/components/ui/button";
-import { FieldValues, SubmitHandler, useForm } from "react-hook-form";
-import { LogInIcon } from "lucide-react";
 import {
+  Button,
   Form,
   FormField,
   FormItem,
   FormLabel,
   FormMessage,
-} from "@/components/ui/form";
+  Input,
+  toast,
+} from "@/components/shadcn";
+import Card from "@/components/ui/Card";
+import { FieldValues, SubmitHandler, useForm } from "react-hook-form";
+import { LogInIcon } from "lucide-react";
+
 import { LogInValues, LogInValuesSchema } from "@/lib/validation";
 import { zodResolver } from "@hookform/resolvers/zod";
-import { PasswordFormField } from "@/components/ui/PasswordFormField";
-import { toast } from "@/components/ui/use-toast";
+import { PasswordFormField } from "@/components/PasswordFormField";
 
 export default function SignInForm() {
   const { isLoaded, signIn, setActive } = useSignIn();
@@ -49,9 +50,9 @@ export default function SignInForm() {
   }, [dirtyFields, passwordValue]);
 
   const disableSubmit =
-  !Boolean(dirtyFields.email) ||
-  !Boolean(dirtyFields.password) ||
-  Boolean(passwordSubmitError);
+    !Boolean(dirtyFields.email) ||
+    !Boolean(dirtyFields.password) ||
+    Boolean(passwordSubmitError);
 
   // Handle the submission of the sign-in form
   const submit: SubmitHandler<FieldValues> = async (data) => {
