+++ src/components/shadcn/calendar.tsx
@@ -5,7 +5,7 @@ import { ChevronLeft, ChevronRight } from "lucide-react";
 import { DayPicker } from "react-day-picker";
 
 import { cn } from "@/lib/utils";
-import { buttonVariants } from "@/components/ui/button";
+import { buttonVariants } from "@/components/shadcn";
 
 export type CalendarProps = React.ComponentProps<typeof DayPicker>;
