+++ src/components/shadcn/form.tsx
@@ -13,7 +13,7 @@ import {
 } from "react-hook-form"
 
 import { cn } from "@/lib/utils"
-import { Label } from "@/components/ui/label"
+import { Label } from "@/components/shadcn"
 
 const Form = FormProvider
