+++ src/components/shadcn/index.ts
@@ -1,0 +1,12 @@
+export * from "./button";
+export * from "./calendar";
+export * from "./dialog";
+export * from "./form";
+export * from "./input-otp";
+export * from "./input";
+export * from "./label";
+export * from "./popover";
+export * from "./textarea";
+export * from "./toast";
+export * from "./toaster";
+export { useToast, toast } from "./use-toast";
