+++ src/components/shadcn/use-toast.ts
@@ -6,7 +6,7 @@ import * as React from "react"
 import type {
   ToastActionElement,
   ToastProps,
-} from "@/components/ui/toast"
+} from "@/components/shadcn"
 
 const TOAST_LIMIT = 1
 const TOAST_REMOVE_DELAY = 1000000
